Update api directory str
