+++ pkg/cmd/config/setAccount.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+	http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -17,6 +17,7 @@ package config
 
 import (
 	"fmt"
+	"github.com/litmuschaos/litmusctl/pkg/apis/experiment"
 	"net/url"
 	"os"
 	"strings"
@@ -158,7 +159,7 @@ var setAccountCmd = &cobra.Command{
 			}
 			utils.White_B.Printf("\naccount.username/%s configured", claims["username"].(string))
 
-			serverResp, err := apis.GetServerVersion(authInput.Endpoint)
+			serverResp, err := experiment.GetServerVersion(authInput.Endpoint)
 			var isCompatible bool
 			if err != nil {
 				utils.Red.Println("\nError: ", err)
