+++ pkg/cmd/create/create.go
@@ -5,7 +5,7 @@ Licensed under the Apache License, Version 2.0 (the "License");
 you may not use this file except in compliance with the License.
 You may obtain a copy of the License at
 
-    http://www.apache.org/licenses/LICENSE-2.0
+	http://www.apache.org/licenses/LICENSE-2.0
 
 Unless required by applicable law or agreed to in writing, software
 distributed under the License is distributed on an "AS IS" BASIS,
@@ -22,14 +22,16 @@ import (
 // createCmd represents the create command
 var CreateCmd = &cobra.Command{
 	Use: "create",
-	Short: `Create resources for LitmusChaos agent plane.
+	Short: `Create resources for LitmusChaos Execution plane.
 		Examples:
 		#create a project
 		litmusctl create project --name new-proj
 
-		#create a Chaos Scenario from a file
-		litmusctl create chaos-scenario -f chaos-scenario.yaml --project-id="d861b650-1549-4574-b2ba-ab754058dd04" --chaos-delegate-id="d861b650-1549-4574-b2ba-ab754058dd04"
+		#create a Chaos Experiment from a file
+		litmusctl create chaos-experiment -f chaos-experiment.yaml --project-id="d861b650-1549-4574-b2ba-ab754058dd04" --chaos-infra-id="d861b650-1549-4574-b2ba-ab754058dd04"
 
+		#create a Chaos Environment
+		litmusctl create chaos-environment --project-id="d861b650-1549-4574-b2ba-ab754058dd04" --name="new-chaos-environment"
 		Note: The default location of the config file is $HOME/.litmusconfig, and can be overridden by a --config flag
 	`,
 }
